+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.0 (WIP)
 
 * Make compatible with alchemy >1.0.7
+* Indicate this datasource does not support ObjectIDs by setting the support flag to false
 
 ## 0.4.0 (2017-08-27)